running on sunshine

+++ Point2D.cpp
@@ -4,7 +4,7 @@ Point2D::~Point2D(void)
 {
 }
 
-Point2D::Point2D(int x, int y)
+Point2D::Point2D(float x, float y)
 {
 	this->x = x;
 	this->y = y;
